Move liquid BIP21 rounding test to liquid module
